openSAFETY Relese Version 1.4
This is the official release of openSAFETY 1.4, as published on SourceForge. It is copied on Github, please refer to the SF repository for the latest iteration.

openSAFETY is an IEC 61508:2010 compliant safety transport protocol suitable for implementing safety-oriented process data transport between individual stations in a safety network. openSAFETY also controls the safety-oriented parameters and configuration of the safety network stations.

The openSAFETY protocol stack was developed by B&R as a reference implementation, in accordance with IEC 61508:2010 of the openSAFETY protocol specification from the EPSG (http://www.ethernet-powerlink.org). Provided as an openSAFETY distribution within the framework of the BSD license together with the documentation for the source code and a CRC calculation tool, it can be downloaded from SourceForge free of charge.

Included as part of the BSD license, an example implementation for openSAFETY 1.4.0 of the safety transport protocol has been made available for demonstration purposes. Without further measures and documentation, this implementation is not suitable for developing products in accordance with IEC 61508:2010. The concrete measures and required documents can be obtained from openSAFETY conformity centers and service providers. A list of certified service providers and test centers for openSAFETY is available from the EPSG.